Add files via upload
This file manages data for creating lists for survey sampling. Organizations send excel files with lists of names, this program takes them in, makes them into a single list, extracting the organization's grant number from the file name, merging with additional information about the organization in another file, checks for duplicates, and reformats the data. 

This type of program is really useful for automating administrative tasks that involve taking groups of files and making a master file where data needs to be reformatted or information needs to be added from another sheet.