Data collection, Data Engineering
This is an interesting program. We are receiving lists of research subjects and their information to take our survey, form a bunch of different organizations. We want to combine these lists into one list, merge with information on a master tracker, and then, assign some new variables based off values merged. 

Finally, we also want to engineer a new variable. This list contains 2 types of research subjects, one primary test subject, and a secondary to be used if each primary declines their consent. For each organization, this program counts the number of those primary test subjects who decline, then marks a new engineered variable as yes, for each secondary test subject until the number of those who declined is reached. The result is, it assigns the secondary replacements to be used as the primary test subjects decline.